+++ src/RcppExports-manual.cpp
@@ -28,23 +28,23 @@ RcppExport SEXP _TreeTools_preorder_edges_and_nodes(SEXP parentSEXP, SEXP childS
   return rcpp_result_gen;
   END_RCPP
 }
-// postorder_edges
-RcppExport SEXP _TreeTools_postorder_edges(SEXP edgeSEXP, SEXP sortSEXP) {
+// postorder_order
+RcppExport SEXP _TreeTools_postorder_order(SEXP edgeSEXP) {
   BEGIN_RCPP
   Rcpp::RObject rcpp_result_gen;
   Rcpp::RNGScope rcpp_rngScope_gen;
   Rcpp::traits::input_parameter< const IntegerMatrix >::type edge(edgeSEXP);
-  Rcpp::traits::input_parameter< const LogicalVector >::type sortSize(sortSEXP);
-  rcpp_result_gen = Rcpp::wrap(TreeTools::postorder_edges(edge, sortSize));
+  rcpp_result_gen = Rcpp::wrap(TreeTools::postorder_order(edge));
   return rcpp_result_gen;
   END_RCPP
 }
 // root_binary
+IntegerMatrix TreeTools::root_binary(const IntegerMatrix edge, const int outgroup);
 RcppExport SEXP _TreeTools_root_binary(SEXP edgeSEXP, SEXP outgroupSEXP) {
   BEGIN_RCPP
   Rcpp::RObject rcpp_result_gen;
   Rcpp::RNGScope rcpp_rngScope_gen;
-  Rcpp::traits::input_parameter< const IntegerMatrix >::type edge(edgeSEXP);
+  Rcpp::traits::input_parameter< Rcpp::IntegerMatrix >::type edge(edgeSEXP);
   Rcpp::traits::input_parameter< const int >::type outgroup(outgroupSEXP);
   rcpp_result_gen = Rcpp::wrap(TreeTools::root_binary(edge, outgroup));
   return rcpp_result_gen;