+++ tests/testthat/test-splits.cpp.R
@@ -1,6 +1,8 @@
 test_that("Bad data results in error", {
-  expect_error(cpp_edge_to_splits(matrix(1, 3, 3), 3))
-  expect_error(cpp_edge_to_splits(matrix(1, 10, 2), 0))
-  expect_error(cpp_edge_to_splits(matrix(1, 10, 2), -10))
-  expect_error(cpp_edge_to_splits(matrix(1, 2, 2), 6)) # Impossible anyway...
+  skip_if(Sys.getenv("USING_ASAN") != "")
+  expect_error(cpp_edge_to_splits(matrix(1, 3, 3), 1:3, 3))
+  expect_error(cpp_edge_to_splits(matrix(1, 10, 2), 1:10, 0))
+  expect_error(cpp_edge_to_splits(matrix(1, 10, 2), 1:10, -10))
+  expect_error(cpp_edge_to_splits(matrix(1, 10, 2), 1:9, 5))
+  expect_error(cpp_edge_to_splits(matrix(1, 2, 2), 1:2, 6)) # Impossible anyway...
 })